rename external tool server
