+++ libs/core/kiln_ai/tools/test_mcp_server_tool.py
@@ -11,7 +11,7 @@ from mcp.types import (
     Tool,
 )
 
-from kiln_ai.datamodel.external_tool import ExternalToolServer, ToolServerType
+from kiln_ai.datamodel.external_tool_server import ExternalToolServer, ToolServerType
 from kiln_ai.tools.mcp_server_tool import MCPServerTool
 from kiln_ai.tools.tool_id import MCP_REMOTE_TOOL_ID_PREFIX
